requirements.txt
# Core
torch>=2.0
torchvision>=0.15
lightning>=2.0
jsonargparse[signatures]>=4.27.7

# Metrics
torchmetrics[detection]>=1.0
faster-coco-eval
pycocotools

# Config
omegaconf

# Utilities
einops
numpy
Pillow
opencv-python

# Logging
loguru
rich
wandb

# Visualization (optional)
graphviz

# HTTP requests (for weight download)
requests
